+++ src/main/java/com/Project/ChickenDeliveryBack/domain/dto/PromotionDetailDTO.java
@@ -1,5 +1,7 @@
 package com.Project.ChickenDeliveryBack.domain.dto;
 
+import java.math.BigDecimal;
+
 public class PromotionDetailDTO {
     private Integer id;
     private Integer idVenta;
@@ -7,8 +9,8 @@ public class PromotionDetailDTO {
     private String codigoPromocion;
     private String nombrePromocion;
     private Integer cantidad;
-    private Double precioUnitario;
-    private Double subtotal;
+    private BigDecimal precioUnitario;
+    private BigDecimal subtotal;
 
     public PromotionDetailDTO() {}
 
@@ -24,8 +26,8 @@ public class PromotionDetailDTO {
     public void setNombrePromocion(String nombrePromocion) { this.nombrePromocion = nombrePromocion; }
     public Integer getCantidad() { return cantidad; }
     public void setCantidad(Integer cantidad) { this.cantidad = cantidad; }
-    public Double getPrecioUnitario() { return precioUnitario; }
-    public void setPrecioUnitario(Double precioUnitario) { this.precioUnitario = precioUnitario; }
-    public Double getSubtotal() { return subtotal; }
-    public void setSubtotal(Double subtotal) { this.subtotal = subtotal; }
+    public BigDecimal getPrecioUnitario() { return precioUnitario; }
+    public void setPrecioUnitario(BigDecimal precioUnitario) { this.precioUnitario = precioUnitario; }
+    public BigDecimal getSubtotal() { return subtotal; }
+    public void setSubtotal(BigDecimal subtotal) { this.subtotal = subtotal; }
 }
